Change Return to Event
This allows for more efficient watching of the blockchain with Web3 (this also is cheaper in terms of gas cost, as there is no storage - just event generation). 

Note, this will require the Dapp front end to be rewritten, but it should simplify things. More can be found here: https://ethereum.stackexchange.com/questions/2024/how-to-access-the-event-log-by-knowing-the-contract-address-web3